refactor cart presenter into lazy array
